mvpっぽく書き直しました。Dialogの仕様が謎だったので一時的にコメントアウトしてます。
