+++ app/src/main/java/com/t_robop/yuusuke/a01_spica_android/BasePresenter.java
@@ -1,0 +1,5 @@
+package com.t_robop.yuusuke.a01_spica_android;
+
+public interface BasePresenter {
+    void start();
+}
