+++ app/src/main/java/com/t_robop/yuusuke/a01_spica_android/BaseView.java
@@ -1,0 +1,5 @@
+package com.t_robop.yuusuke.a01_spica_android;
+
+public interface BaseView<T> {
+    void setPresenter(T presenter);
+}
