+++ app/src/main/java/com/t_robop/yuusuke/a01_spica_android/Script/ScriptActivity.java
@@ -1,0 +1,31 @@
+package com.t_robop.yuusuke.a01_spica_android.Script;
+
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentTransaction;
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+import com.t_robop.yuusuke.a01_spica_android.R;
+
+public class ScriptActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_script);
+
+        // Fragmentを作成します
+        Fragment fragment = new ScriptFragment();
+        // Fragmentの追加や削除といった変更を行う際は、Transactionを利用します
+        FragmentTransaction transaction = getSupportFragmentManager().beginTransaction();
+        // 新しく追加を行うのでaddを使用します
+        // 他にも、よく使う操作で、replace removeといったメソッドがあります
+        // メソッドの1つ目の引数は対象のViewGroupのID、2つ目の引数は追加するfragment
+        transaction.add(R.id.container, fragment);
+        // 最後にcommitを使用することで変更を反映します
+        transaction.commit();
+
+        new ScriptPresenter((ScriptContract.View) fragment);
+    }
+
+}
