+++ app/src/main/res/values/strings.xml
@@ -18,4 +18,7 @@
         <item>0003</item>
         <item>0004</item>
     </string-array>
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
